third_party/WebKit/LayoutTests/virtual/color_space/fast/canvas/color-space/canvas-colorSpace-attribute.html

<!DOCTYPE HTML>
<script src="../../../../../resources/testharness.js"></script>
<script src="../../../../../resources/testharnessreport.js"></script>
<script>
function checkCanvasColorSpace(requestedColorSpace, expectedColorSpace) {
	var canvas = document.createElement("canvas")
	var ctx = canvas.getContext('2d', { colorSpace: requestedColorSpace });
	assert_equals(ctx.getContextAttributes().colorSpace, expectedColorSpace);
}

generate_tests(checkCanvasColorSpace, [
	['legacy-srgb is supported', 'legacy-srgb', 'legacy-srgb'],
	['srgb is supported', 'srgb', 'srgb'],
	['linear-rgb is supported', 'linear-rgb', 'linear-rgb'],
	['rec-2020 is supported', 'rec-2020', 'rec-2020'],
	['p3 is supported', 'p3', 'p3'],
]);
</script>
